Fix step-1 failure for boundary states on the velocity-at-zero-acceleration curve
Reported input (1 DoF, v0 == a0^2/(2*j_max) and vf == af^2/(2*j_max) exactly):
step 1 found four valid profiles - the NONE-family profile and a degenerate
VEL-family profile (zero-width velocity plateau) with identical duration and
the same direction. Block::calculate_block only merged near-identical
profiles with opposite directions, so the even profile count made the
calculation fail, while any 1e-12 perturbation of the input succeeded.

Two changes to the duplicate handling in Block::calculate_block:
- The (0,3) profile pair no longer requires opposite directions, matching
  upstream C++ ruckig behavior (the direction condition was added in the
  port by mistake)
- Fallback for the remaining even-count cases: scan all pairs and merge two
  profiles whose durations differ by less than 1e-12*(1+t) - such a pair
  describes the same solution and the interval between them is degenerate
  (this also fixes neighboring inputs on the other side of the degeneracy
  that failed with the pair (1,3))

Adds a regression test with the reported input and six perturbations around
the degeneracy. Also moves the tracking allocation-count test into its own
test binary, as the global allocation counter is disturbed by other tests
running in parallel threads of the same process.
